Harden version checker: block-scoped regex, split files, assert matches
- Use regex with whitespace capture for replacements, preventing gsub
  collisions when versions share field values
- Split VersionChecker, RubyChecker, NodeChecker into separate files
- apply! raises on unmatched replacements instead of silent no-op
- Fix unintended ### Added header in CHANGELOG
